Use `IdentityHashMap` as the underlying storage for `SoyRecord`, switch to a new key type (`RecordSymbol`) that has intrinsic identity semantics.
This helps because IdentityHashMap has a number of advantages over ImmutableMap for `record` usecases
1. it avoids allocating `Map.Entry` objects until iterated, and we mostly do not iterate.
   - When we do we can use the `forEach` method to avoid allocation overhead
2. It uses an open addressing scheme which flattens the memory layout
3. relying on System.identityHashCode and `==` comparisons is faster than String.hasCode and String.equals

The downside of course is that you need keys with _identity_ semantics.  Because all the keys for record lookups are `String` literals this was mostly already true (plugin implementations being the exemption) but due to ambiguity and the general oddness of using identity semantics for strings, I have introduced a new type `RecordProperty` to clarify semantics.

PiperOrigin-RevId: 573626909