Remove a layer of indirection from template params by refactoring `ParamStore` to be a subclass of `IdentityHashMap`
Furthermore use `ParamStore` exclusively instead of `SoyRecord` for template parameters and as the internal 'backing store' of `SoyRecordImpl`.

This should improve performance by:
1. Removing indirection from parameter access
2. Decreasing the number of objects that need to be allocated
3. Making all parameter access 'mono-morphic' (`ParamStore` is a `final` class)

This also makes things a little bit safer by allowing us to `freeze` the ParamStore to verify mutation semantics.

The only particular cost (beyond churn) is that `data="$expr"` calls get a bit more expensive, since we will need to adapt a `SoyRecord` to a `ParamStore`.  Because `SoyRecordImpl` wraps a `ParamStore` we can provide a fastpath there, but for `record` values passed through the generic `SoyValueConverter` path we will induce a copy.

PiperOrigin-RevId: 574198561