Merge `MethodRef` and `ConstructorRef`
There is only a slight difference between these two types due to slightly differing calling conventions but this is easy enough to handle in `MethodRef` directly with a few if-statements.

This also allows constructors to trivially take advantage of the constant folding support added to `MethodRef.invoke`

Finally move the static constants out of `MethodRef` to a new `MethodRefs` class to clarify implementation vs uses.

PiperOrigin-RevId: 575625659