Extract a helper class for constructing lambdas from TemplateCompiler
This will soon be reused in a few different places.

Add a tiny optimization for lambdas with zero bound parameters by compiling them to an `ldc` instruction.  This special case happens to satisfy all of our current usecases and will satisfy a subset of the upcoming ones.  This should be a tiny bootstraping and codesize improvement.

PiperOrigin-RevId: 575868161