Optimize the standard case of a translating a with no placeholders
 * eliminate the allocation of singleton immutable lists
 * eliminate the SoyMsgRawTextPart

This reduces allocations (static for the default and dynamic when doing lookups) and reduces some indirection.

Also optimize the `escapeHtml` routine for the fact that our search character is a single ascii char. jvms have optimized vm intrinsics for `indexOf(char)` and we can leverage that here.

And finally add a dedicated `hasMsg` operation to improve readability and save allocations in the common case.

PiperOrigin-RevId: 592690615