Move `ijData` into `RenderContext`
Both ijData and RenderContext are constant across the render tree are constant across the render tree so we can save some overhead by combining them.

Note, the output buffer cannot be combined in this way since it is routinely wrapped for streaming escaping and thus changes across the render tree.

From this we should expect a small code size and performance win.  The performance win should result from:

* fewer objects to pass/capture in lazy closures
* decreased register pressure

PiperOrigin-RevId: 597062534