Use a custom interface for passing ijdata instead of a `Map`
A complex part of setting up the render context is preparing ijdata, to assist we provide methods like `getTransitiveIjParamsForTemplateRender` that allow callers to figure out the super set of required params and make sure they are supplied.

This is wasteful for a few reasons
* calls like `getTransitiveIjParamsForTemplateRender` are simply expensive to implement and execute
  - also because it does not recurse through all datastructures it may miss some templates.
* `getTransitiveIjParamsForTemplateRender` can only compute the super-set of required parameters, not the actual set, so generally callers are passing more than what is required.

To address this we replace the map with a simple functional interface.  This will allow callers to compute ij parameters on demand as they are required by execution.

This change will not actually change behavior as the behavior change requires callers to pass something new to SoySauce.

PiperOrigin-RevId: 597582821