Update to [Truth 1.4.0](https://github.com/google/truth/releases/tag/v1.4.0).
This makes available the rest of the Truth APIs that were recently added in Google's monorepo.

It may be worth also updating to the _following_ version of Truth after it's released, but the purpose of that release will mostly be to deprecate `Truth8`. So, if you just avoid using `Truth8` now (starting by approving any migration CLs that I send your way), then there will be little need for that upgrade. Or, more to the point: I don't expect to keep sending you weekly upgrade CLs after this one :)

PiperOrigin-RevId: 604706771