Improve support for loops and collection access in `TemplateAnalysis`
Some collection expressions in soy are guaranteed to never include futures, notably anything returned from a proto or various builtin functions/methods syntax structures, we can take advantage of that to eliminate runtime doneness checks.  This generalizes the logic for ranges to apply to things like

```
{for $v in $proto.getFooList()}
  {$v} // doesn't need to be checked.
{/for}
```

This is accomplished by analyzing the source of values, if they come from a known resolved source, then we know that accessing them will return a resolved value.

PiperOrigin-RevId: 632631730