Reimplement `let` and `param` values as `static` methods + invokedynamic generated classes instead of inner classes.
This will unlock optimistic evaluation in a followup change.  In the mean time this will be a massive decrease in the number of classes generated and the overall size of those classes since we eliminate a ton of 'constant pool duplication' and basic overheads.  We expect a small runtime regression due to the extra layer of indirection from the new static methods.

One tricky thing was reworking the naming algorithm for these 'inner methods'. For that we are using a hierarchical naming scheme where the name of the methods are a combination of the parent template and the local param/let variable.

PiperOrigin-RevId: 635474405