Inline the RecordToPositionalCallFactory.  It turns out there is some overhead to the methodhandle delegate
This was used to generate the `(ParamStore,LoggingAdvisingAppendable,RenderContext)RenderResult` signature function for templates with positional parameters.  This signature is used when called directly from `SoySauceImpl` and when the template is invoked as a template variable or with a `data=` expression.  This should slightly reduce overhead of top level calls.

PiperOrigin-RevId: 638002270