Allow `const` and `extern` methods to be defined without a `RenderContext` parameter when it is unneeded by the implementation method.
This is a small code size optimization and provides a hint for 'constantdyanmic' based constant folding.  When we 'link' to a `const` definition without a `RenderContext` parameter, we can just inline the constant value into the callsite.  Similarly when calling externs we can just drop the `RenderContext` parameter which should improve generated code.

For calls to these methods from the same file, we can do these optimizations at compile time, but for cross file calls we rely on our dynamic linking via `invokedynamic` which can generate an 'ad hoc' adapter to target the correct signature.

This is a small code size and performance optimization.

PiperOrigin-RevId: 638386423