Detect when our lazy closures require no detaches and evaluate them "inline" instead of capturing into SoyValueProvider subclasses.
On its own this is smallish performance optimization and it critically lays some groundwork for optimistic evaluation. Because we are changing evaluation order some things that may previously have been lazily evaluated (and thus possibly not evaluated at all), now will be eagerly evaluated. To mitigate risk from hidden errors being reported we use a try-catch block to capture exceptions thrown during eager evaluation and defer them until the end of rendering.

In theory, every time this optimization fires, it is demonstrating an issue with the static analysis (e.g. `TemplateAnalysis`) in other parts of the compiler.  Fully resolving all of that is likely to be much more complex, instead we can probably reduce the reliance on static analysis from other parts of the compiler and be more 'reactive' like this instead.

PiperOrigin-RevId: 641276680