Change where `setKindAndDirectionality` is called to reduce how often we need to call it.
This method informs the current output appendable what kind of content is currently being printed which in turn allows for certain escaping directives to become no-ops.  Currently we output calls to this method at the top of every template/let/param.  To reduce the number of calls we can leverage the static analysis already performed by the autoescaper.  Our new strategy is to only call the method when the content kind _might_ be changing.  Which is:

* top level calls (java->soy)
* dynamic calls (`{call $expr /}`)
* when content kind is changing (a subset of soy->soy)

For the first two cases the call always goes through the `CompiledTemplate` interface, so this cl creates a new factory for those class so we can insert the `setKindAndDirectionality` calls see `CompiledTemplateMetaFactory`

For the last case we can just modify the code for `call` expressions to insert calls to `setKindAndDirectionality` when we are **not** calling through a `CompiledTemplate` object and when we are allocating new streaming escapers.  If we are not allocating a new appendable then we know the content kind is not changing.

Because that last case is by far the most common case we can expect this to be a performance improvement.

Finally, in the case of a dynamic call, we add support for wrapping appendables to cut themselves out of the call stack. This only applies to a small number of escapers, but it ensures that if we do a dynamic call to a safe html template from a safe html context, we don't need to call through the EscapeHtml delegate appendable.

PiperOrigin-RevId: 644538555