Add support for parsing `SanitizedContent` of `kind=attributes` into a `Map<String, AttributeValue>`
Also support a restricted mechanism to build a SanitizedContent object from a map.

This is useful for soy externs that need to manipulate a set of attributes in a non-trivial way.

In a followup additional support will be added to the code generator to support producing these directly, however the 'parse from string' path will still be needed to support users directly constructing attribute values.

PiperOrigin-RevId: 645506497