Support capturing rendering output into a `SoyValue` without dropping logs.
This requires changing the representation of `SanitizedContent` and `StringData` to optionally store `List<logging commands>` and changing the compiler to call `SoyValue.render(appendable)` instead of `appendable.append(value.coerceToString())` when printing a boxed soy value.

Interestingly, this solution was discussed in the original design of logging, however it was abandoned in favor of streaming, which does mostly address the same concern.  However, because it failed to address the complete issue of dropped logs we still need to do this.

Unfortunately, the change to call `render` is a performance regression, simply due to virtual method overhead afaict.  For most call sites `appendable.append(value.coerceToString())` both the `append` call and the `coerceToString()` call are monomorphic and thus can be easily optimized.  However, for a call like `value.render(appendable)`, the `render` method is monomorphic but the `append` call that is inside of the `render` call is **not**.  The regression is minor however and can easily be recovered in other ways.

PiperOrigin-RevId: 646107161