Optimistically evaluate ATTRIBUTES blocks
After further analysis of the brittle soy externs it is clear that `kind=attributes` blocks are not affected and so don't need to be held back.

This should be a performance enhancement as now we will only allocate `DetachableContentProvider` subtypes for attributes blocks when we need to evaluate an unresolved future.  This will save runtime, allocations and classloading.

PiperOrigin-RevId: 648747211