Simplify velog gencode
* Eliminate `hasLogger` checks
* Don't explicitly pass `null` to ve_data (do this in js too)
* Try to constant fold `ve_data` calls and `LogStatement.create`
* Eliminate the `SoyLogger.NO_OP` api and instead use `null` for signaling.

The `hasLogger` tests in the gencode for jbcsrc were an attempt to decrease the cost of adding `{velog}` commands to shared libraries even though only a subset of callers would ever install a logger.  This turned out to be an expensive and unnecessary optimization (common shared libraries never got instrumeted).  With this approach we will always execute logging commands, but constructing `LogStatement` values should be cheap and could be optimized anyway (why not pass `SoyVisualElementData` to the logger directly?), if needed.

PiperOrigin-RevId: 650704247