Use a hash table for looking up messages instead of a hash+binary search approach.
Historically we avoided a hash table to avoid the overheads from open addressing (empty slots), or closed addressing (entry object chains). As well as the overheads of boxed `Long` keys. Instead the hash+ranged binary search implemented a custom closed addressing system which was clever but this required a parallel `int[]` to store ranges, and the binary search required additional comparisons/reads.  With a more traditional hashtable approach we do absorb the overheads of hashing but since we expect the structure of all `RenderOnlySoyMsgBundle` objects to be similar we can amortize the overhead by simply sharing the table across all bundles.

The implementation is a straightforward open address hash table of `long->int` where the `int` values are the location where the message should be stored inside the bundle array.  This means that all `RenderOnlySoyMsgBundle` objects will have the same memory layout.

PiperOrigin-RevId: 658181040