Simplify the gencode for idom ve logging when the `logOnly` parameter is used.
Move the logic of whether to construct a `NullRenderer` into the `enter` functions, now the gencode just needs to capture the return value when passing a logonly parameter.

This saves 2 `if` statements and one local variable per `velog` with a logonly parameter.  This also fixes a more basic issue where we would accidentally generate the logic for the logonly expression twice!.

Clean up a few minor issues at the same time:
 * use `undefined` instead of `null` to save time and make dead property analysis simpler
 * rename `enter`/`exit` to disambiguate them.

PiperOrigin-RevId: 707130737