New API for OutputVarHanlder to support lazy execution. This is a no-op.
Each OutputVar is associated with a `Style` that will do string concatenation or buffer methods. The default is APPENDING, which is the existing behavior.

The initialized bit is now a tristate to support declaring a variable without initializing it from outside of a block.

To allow chaining, we'll have a new API to add Expressions grouped together into `Parts`, each of which corresponds to a single add() call.

PiperOrigin-RevId: 753273683